course-files/lectures/lecture25/requirements.txt
requests==2.28.2        # for testing
dnspython==2.3.0        # support for connecting to remote MongoDB
python-dotenv==1.0.0    # support for environment variables
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.4.4

psycopg2-binary==2.9.5  # PostgreSQL DB connector
SQLAlchemy==2.0.7
Flask-SQLAlchemy==3.0.3 # Python <--> DB ORM
pylint-django==2.5.3    # linter help
Faker==18.3.1           # to install fake data
gunicorn==20.1.0
